print dependencies in a stable order
When a service has multiple direct dependencies, the order in which they
were printed to the console different from execution to execution.

We need to produce a stable reproducible dependency-list to be able to
rerun the tool and detect if the dependencies changed or not.

To achieve that dependencies that are on the same level in the
dependency tree are now printed in ascending alphabetically order.
This required:
- the nodeps list to be sorted to always append it to the dependency
  list in the same order,
- change the Graph.TopologicalSort() method to iterate over the Vertices
  always in the same order.
  To be able to do that the required code from the vendored
  thcyron/graphs package was copied to a local package, the Vertex type
  was replaced with a string and the TopologicalSort() method was
  adapted.
  This adaption is hacky and less efficient then the original one, but
  it is sufficient for our humble use case.

  [PLAT-5498]