remove -docker-compose parameter
we don't use the parameter, remove support for it.
This also allowed to simplify the Service struct.
The Service.DependsOn map value was DepService struct.
The DepService struct that was used only when unmarshaling a
docker-compose file. It was not used in other circumstances and also
never read.

The DepService struct is removed and the value type of Service.DependsOn
is changed to an empty struct.