cli: refactor command line interface, introduce sub commands
Refactor the command line interface to use sub commands for operations,
instead of parameters.
The popular cobra package is used to manage the CLI.
Parameters now follow the GNU convention for long (--) and short (-)
parameters.
The required repository path, env and region parameters are converted to
positional arguments from parameters.

The cobra commands are defined in the new internal/cmd package.
It follows the recommended layout[^1].

The main.go file now only runs cmd.Execute().
The logic for the commands is moved to the Run methods of the
sub commands defined in the cmd package.

The code was before in the main package (except the main func) is mvoed
to internal/deps. This was necessary to be able to import it from
internal/cmd.

[^1]: https://github.com/spf13/cobra/blob/main/site/content/user_guide.md