support ignoring undeployable apps
We have some apps that can not be deployed but have dependencies
defined, e.g. the meta-apps.
Those apps are included in the generated deployment orders.
In multiple places we have code snippets to remove these again from the
deploy-order.
To get rid of those snippets, support to figure out if an app is
deployable and store the information for it in the dependency-tree.

An app is marked as undeployable if it does not have a deploy directory.

The deployorder command is changed to not include undeployable apps by
default. This can be changed by passing the new --include-undeployable
parameter.