deploy-order: fail if non-existing app is specified
If an application was specified to calculate the deploy-order for that
did not exist in the tree, the app was added to the tree as
undeployable, eventually included in the deploy order and the command
succeed.

Fail instead if the app does not exist in the dependency tree.
This will allow to detect during deployment if an app was specified that
does not exist in the exported dependency tree