export: make DEST-FILE argument optional, write to stdout by default (#23)
Change the export command to write the result by default to stdout and
make the DEST-FILE parameter optional.

To simplify the code a bit, change "export [DEST-FILE]" also to
print the cleaned path instead of the absolute path of DEST-FILE.
This is sufficient. The caller knows where the file is he writes the
result to.